fix: BoardService like 합치는 중

+++ src/main/java/kr/ac/soup/service/BoardServiceImpl.java
@@ -9,12 +9,14 @@ import kr.ac.soup.entity.Member;
 import kr.ac.soup.repository.BoardRepository;
 import kr.ac.soup.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.management.InstanceNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -37,8 +39,11 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     @Transactional // Proxy를 초기화시킬려면 Repository의 Transaction을 끌고와야 함
-    public BoardResponseDto getBoard(Long boardId) {
+    public BoardResponseDto getBoard(Long boardId) throws InstanceNotFoundException {
         Optional<Board> findBoard = boardRepository.findById(boardId);
+        if(findBoard.isEmpty()){
+            throw new InstanceNotFoundException("게시글을 찾지 못하였습니다.");
+        }
         Board board = findBoard.get();
 
         List<ReplyResponseDto> repliesDto = new ArrayList<>();
@@ -65,10 +70,10 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public Long postBoard(BoardPostRequestDto boardPostRequestDto) {
+    public Long postBoard(BoardPostRequestDto boardPostRequestDto) throws InstanceNotFoundException {
         Optional<Member> findMember = memberRepository.findById(boardPostRequestDto.getMemberId());
         if (findMember.isEmpty()) {
-            return null;
+            throw new InstanceNotFoundException("유저를 찾지 못하였습니다.");
         }
         Board board = Board.builder()
                 .member(findMember.get())
@@ -83,10 +88,10 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public Long updateBoard(Long boardId, BoardPostRequestDto boardPostRequestDto) {
+    public Long updateBoard(Long boardId, BoardPostRequestDto boardPostRequestDto) throws InstanceNotFoundException {
         Optional<Board> findBoard = boardRepository.findById(boardId);
         if (findBoard.isEmpty()) {
-            return null;
+            throw new InstanceNotFoundException("게시글을 찾지 못하였습니다.");
         }
 
         Board board = findBoard.get();
@@ -95,8 +100,11 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public Long deleteBoard(Long boardId) {
+    public Long deleteBoard(Long boardId) throws InstanceNotFoundException {
         Optional<Board> findBoard = boardRepository.findById(boardId);
+        if(findBoard.isEmpty()){
+            throw new InstanceNotFoundException("게시글을 찾지 못하였습니다.");
+        }
         Board board = findBoard.get();
         board.deleteBoard();
         boardRepository.delete(board);
